final-project-kem-entertainment-main/requirements.txt
streamlit
vertexai
bigquery
google-generativeai
pytest
mock


